Autocompleter label row: fix margin-driven misalignment; merge stray parity test file
Bootstrap's label { margin-bottom: 5px; } sits on only one side of
the label's margin box, throwing off FieldLabelRow's align-items-center
centering against between-content (has-id-indicator, find/keep/edit
icons) that carries no margin of its own. Zeroed the label's margin
inside .autocompleter and moved the same bottom spacing onto the
whole label row instead (new .autocompleter-label-row class), so only
the source of the spacing changes, not the total spacing below the
row.

application_form_helper_parity_test.rb had no ERB counterpart left to
have parity with (the ERB migration is complete) and tested the exact
same ApplicationForm field helpers as application_form_test.rb, just
with a different fixture-class-per-test style instead of the
established render_form { ... } block pattern. Converted its ~19
fixture form classes to that pattern and merged all tests into
application_form_test.rb; deleted the standalone file and fixed the
one stray comment reference to it.
